Updated home page to a more simple design
Removed all of the promotional and content that explains perdiem. The
reason for this is to make it feel more “cool” and “exclusive”. Like
something that people are part of that no one else knows about.

This has been one of the biggest reasons for our success in other
areas. Applying it to here I think will do the same. I’m working on a
better “how-it-works” section that we can link at the bottom with
artist and investor resources.

Also think it would be great for once you are logged in to have the
default go to your user profile. This way it feels more like you are
inside of the system and want to use and explore it more. Wasn’t
exactly sure how to do this.