Updated leaderboards formatting, added comparison chart
I really wanted to do a lot more with these but couldn’t figure out how
to get the list to tie into the charts.js.

I was able to put a pie chart in that breaks down the raised vs
invested.

Would like to be able to have each member be ranked based on % earned
rather than total amount earned. I was having trouble with figuring out
how to make this happen.

I was also thinking it would be great to have them ranked on a chart
based on amount or percentage. Still need to put some thought into this
but wanted to see if there was an easier way to pull this data