Updated registration page to encourage oauth
Most people want to login with oath but don’t know that it’s possible
when they see the big boxes to fill out.

I made the manual process into a button that pops up into a modal. I
think this might be a lot better for people to know it’s possible to
login with Facebook or google