Removed pie chart from leaderboard, added {% block extras %}
Would like to talk more about this section. I tried for days to get the
leaderboard data to display in a chart but couldn’t get any of the data
to load properly. I’m curious to know if this is a backend thing.

Would like this section to be more visual. For example a bar chart with
all of the top leaders and how much they have earned / generated vs
invested / raised.