Removed copy from register page
One note on this page: When a new user signs up it doesn’t sign them up
for artist updates. They have to go in to settings and subscribe to
them manually. This was a huge problem because no one was getting
updates from the artists

This caused the artists to not want to post updates because no one was
receiving them. Everyone both artists and investors both really want
the “subscribe to updates from artists” selected by default.