Fixed formatting on artist updates
Tried figuring out how to get the text input field larger, couldn’t get
it to work. I think it’s something happening on the back-end.

Everyone has requested to have the input field bigger. It’s hard to
read when trying to write stuff in the box.

The biggest request is being able to edit the post. Not sure how
difficult it is to make them editable, but that would be amazing to
have if possible.