Updated invest and past campaign page formatting
Updated static image to dynamic artwork. Would be great if this could
show the actual project artwork, but not essential.

Added more details to number of share and total cost

Removed “Each Share is Worth” - This is the section that I we needed to
update the JS for. I missed it in the last PR.  {{ campaign.share_price
}}.