Fixed charts on music stats page
Everything should be back to working normally. Spotify just did a
significant change to their estimate system and it’s no longer publicly
available. We should probably switch the “estimated revenue” to
“revenue generated” eventually

Would be great if we could have the data for streams and downloads
display monthly the same way that estimated revenue does. This isn’t
essential, but if it’s an easy fix then would be nice to have